TeX {plotly}	R Documentation

Render TeX in a plotly graph using MathJax

Description

This function makes it slightly easier to render TeX in a plotly graph – it ensures that MathJax is included with the final result and also ensures the provided string is surrounded with $ (this is what plotly.js uses to declare a string as TeX).

Usage

TeX(x)

Arguments

x
a character vector

See Also

config

Examples

plot_ly(x = c(1, 2, 3, 4), y = c(1, 4, 9, 16)) %>%
  layout(title = TeX("\\text{Some mathjax: }\\alpha+\\beta x")) %>%
  config(mathjax = "cdn")

[Package plotly version 4.10.3 Index]