Separate groups with missing values

This function is used internally by plotly, but may also be useful to some power users. The details section explains when and why this function is useful.

group2NA(
  data,
  groupNames = "group",
  nested = NULL,
  ordered = NULL,
  retrace.first = inherits(data, "GeomPolygon")
)

Arguments

data
a data frame.

groupNames
character vector of grouping variable(s)

nested
other variables that group should be nested (i.e., ordered) within.

ordered
a variable to arrange by (within nested & groupNames). This is useful primarily for ordering by x

retrace.first
should the first row of each group be appended to the last row? This is useful for enclosing polygons with lines.

Value

a data.frame with rows ordered by: nested, then groupNames, then ordered. As long as groupNames contains valid variable names, new rows will also be inserted to separate the groups.

Details

If a group of scatter traces share the same non-positional characteristics (i.e., color, fill, etc), it is more efficient to draw them as a single trace with missing values that separate the groups (instead of multiple traces), In this case, one should also take care to make sure connectgaps is set to FALSE.

Examples

# note the insertion of new rows with missing values 
group2NA(mtcars, "vs", "cyl")

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 2    NA   4    NA  NA   NA    NA    NA  0 NA   NA   NA
#> 3  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 4  24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 5  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 6  32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 7  30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 8  33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 9  21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 10 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 11 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 12 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 13 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 14 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 15 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 16   NA   6    NA  NA   NA    NA    NA  0 NA   NA   NA
#> 17 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 18 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 19 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 20 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 21 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 22 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 23 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 24 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 25 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 26 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 27 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 28 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 29 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 30 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 31 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 32 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 33 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 34 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8

# need to group lines by city somehow!
plot_ly(txhousing, x = ~date, y = ~median) %>% add_lines()

# instead of using group_by(), you could use group2NA()
tx <- group2NA(txhousing, "city")
plot_ly(tx, x = ~date, y = ~median) %>% add_lines()

# add_lines() will ensure paths are sorted by x, but this is equivalent
tx <- group2NA(txhousing, "city", ordered = "date")
plot_ly(tx, x = ~date, y = ~median) %>% add_paths()