highlight_key {plotly}	R Documentation

Highlight/query data based on primary key

Description

This function simply creates an object of class crosstalk::SharedData. The reason it exists is to make it easier to teach others how to leverage its functionality in plotly. It also makes it more discoverable if one is already aware of highlight.

Usage

highlight_key(x, ...)

Arguments

x
a plotly visualization or a data.frame.

...
arguments passed to crosstalk::SharedData$new()

Value

An object of class crosstalk::SharedData

Author(s)

Carson Sievert

See Also

highlight

[Package plotly version 4.10.3 Index]