knit_print.api_grid_local {plotly}	R Documentation

Embed a plotly grid as an iframe in a knitr doc

Description

Embed a plotly grid as an iframe in a knitr doc

Usage

knit_print.api_grid_local(x, options, ...)

Arguments

x
a plotly figure object

options
knitr options.

...
placeholder.

References

https://github.com/yihui/knitr/blob/master/vignettes/knit_print.Rmd

[Package plotly version 4.10.3 Index]