last_plot {plotly}	R Documentation

Retrieve the last plot to be modified or created.

Description

Retrieve the last plot to be modified or created.

Usage

last_plot()

See Also

ggplot2::last_plot()

[Package plotly version 4.10.0 Index]