plotly-shiny {plotly}	R Documentation

Shiny bindings for plotly

Description

Output and render functions for using plotly within Shiny applications and interactive Rmd documents.

Usage

plotlyOutput(
  outputId,
  width = "100%",
  height = "400px",
  inline = FALSE,
  reportTheme = TRUE,
  fill = !inline
)

renderPlotly(expr, env = parent.frame(), quoted = FALSE)

Arguments

outputId
output variable to read from

width, height
Must be a valid CSS unit (like "100%", "400px", "auto") or a number, which will be coerced to a string and have "px" appended. Note that, for height, using "auto" or "100%" generally will not work as expected, because of how height is computed with HTML/CSS.

inline
use an inline (span()) or block container (div()) for the output

reportTheme
whether or not to report CSS styles (if a sufficient version of shiny and htmlwidgets is available).

fill
see htmlwidgets::shinyWidgetOutput() for explanation (requires a recent version of htmlwidgets).

expr
An expression that generates a plotly

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

[Package plotly version 4.10.3 Index]