plotly_example {plotly}	R Documentation

Run a plotly example(s)

Description

Provides a unified interface for running demos, shiny apps, and Rmd documents which are bundled with the package.

Usage

plotly_example(type = c("demo", "shiny", "rmd"), name, edit = TRUE, ...)

Arguments

type
the type of example

name
the name of the example (valid names depend on type).

edit
whether to open the relevant source files using file.edit. Only relevant if type is "shiny" or "rmd".

...
arguments passed onto the suitable method.

Author(s)

Carson Sievert

[Package plotly version 4.10.3 Index]