raster2uri {plotly}	R Documentation

Encode a raster object as a data URI

Description

Encode a raster object as a data URI, which is suitable for use with layout() images. This is especially convenient for embedding raster images on a plot in a self-contained fashion (i.e., so they don't depend on external URL links).

Usage

raster2uri(r, ...)

Arguments

r
an object coercable to a raster object via as.raster()

...
arguments passed onto as.raster().

Author(s)

Carson Sievert

References

https://plotly-r.com/embedding-images.html

Examples

# a red gradient (from ?as.raster)
r <- as.raster(matrix(hcl(0, 80, seq(50, 80, 10)), nrow = 4, ncol = 5))
plot(r)

# embed the raster as an image
plot_ly(x = 1, y = 1) %>% 
  layout(
    images = list(list(
     source = raster2uri(r),
     xref = "paper", 
     yref = "paper", 
     x = 0, y = 0, 
     sizex = 0.5, sizey = 0.5, 
     xanchor = "left", yanchor = "bottom"
  ))
 ) 

[Package plotly version 4.10.3 Index]