colorbar {plotly}	R Documentation

Modify the colorbar

Description

Modify the colorbar

Usage

colorbar(p, ..., limits = NULL, which = 1)

Arguments

p
a plotly object

...
arguments are documented here https://plotly.com/r/reference/#scatter-marker-colorbar.

limits
numeric vector of length 2. Set the extent of the colorbar scale.

which
colorbar to modify? Should only be relevant for subplots with multiple colorbars.

Author(s)

Carson Sievert

Examples

p <- plot_ly(mtcars, x = ~wt, y = ~mpg, color = ~cyl)

# pass any colorbar attribute -- 
# https://plotly.com/r/reference/#scatter-marker-colorbar
colorbar(p, len = 0.5)

# Expand the limits of the colorbar
colorbar(p, limits = c(0, 20))
# values outside the colorbar limits are considered "missing"
colorbar(p, limits = c(5, 6))

# also works on colorbars generated via a z value
corr <- cor(diamonds[vapply(diamonds, is.numeric, logical(1))])
plot_ly(x = rownames(corr), y = colnames(corr), z = corr) %>%
 add_heatmap() %>%
 colorbar(limits = c(-1, 1))

[Package plotly version 4.10.0 Index]