embed_notebook {plotly}	R Documentation

Embed a plot as an iframe into a Jupyter Notebook

Description

Embed a plot as an iframe into a Jupyter Notebook

Usage

embed_notebook(x, width = NULL, height = NULL, file = NULL)

Arguments

x
a plotly object

width
attribute of the iframe. If NULL, the width in plot_ly is used. If that is also NULL, '100%' is the default.

height
attribute of the iframe. If NULL, the height in plot_ly is used. If that is also NULL, '400px' is the default.

file
deprecated.

Author(s)

Carson Sievert

[Package plotly version 4.10.0 Index]