event_data {plotly}	R Documentation

Access plotly user input event data in shiny

Description

This function must be called within a reactive shiny context.

Usage

event_data(
  event = c("plotly_hover", "plotly_unhover", "plotly_click", "plotly_doubleclick",
    "plotly_selected", "plotly_selecting", "plotly_brushed", "plotly_brushing",
    "plotly_deselect", "plotly_relayout", "plotly_restyle", "plotly_legendclick",
    "plotly_legenddoubleclick", "plotly_clickannotation", "plotly_afterplot",
    "plotly_sunburstclick"),
  source = "A",
  session = shiny::getDefaultReactiveDomain(),
  priority = c("input", "event")
)

Arguments

event
The type of plotly event. All supported events are listed in the function signature above (i.e., the usage section).

source
a character string of length 1. Match the value of this string with the source argument in plot_ly() (or ggplotly()) to respond to events emitted from that specific plot.

session
a shiny session object (the default should almost always be used).

priority
the priority of the corresponding shiny input value. If equal to "event", then event_data() always triggers re-execution, instead of re-executing only when the relevant shiny input value changes (the default).

Author(s)

Carson Sievert

References

https://plotly-r.com/linking-views-with-shiny.html#shiny-plotly-inputs

https://plotly.com/javascript/plotlyjs-function-reference/

See Also

event_register, event_unregister

Examples

## Not run: 
plotly_example("shiny", "event_data")

## End(Not run)

[Package plotly version 4.10.0 Index]