event_data {plotly}	R Documentation

Access plotly user input event data in shiny

Description

This function must be called within a reactive shiny context.

Usage

event_data(event = c("plotly_hover", "plotly_click", "plotly_selected",
  "plotly_relayout"), source = "A",
  session = shiny::getDefaultReactiveDomain())

Arguments

event
The type of plotly event. Currently 'plotly_hover', 'plotly_click', 'plotly_selected', and 'plotly_relayout' are supported.

source
a character string of length 1. Match the value of this string with the source argument in plot_ly() to retrieve the event data corresponding to a specific plot (shiny apps can have multiple plots).

session
a shiny session object (the default should almost always be used).

Author(s)

Carson Sievert

Examples

## Not run: 
plotly_example("shiny", "event_data")

## End(Not run)

[Package plotly version 4.8.0 Index]