export {plotly}	R Documentation

Export a plotly graph to a static file

Description

This function is in the process of being deprecated (use orca instead).

Usage

export(p = last_plot(), file = "plotly.png", selenium = NULL, ...)

Arguments

p
a plotly or ggplot object.

file
a filename. The file type is inferred from the file extension. Valid extensions include 'jpeg' | 'png' | 'webp' | 'svg' | 'pdf'

selenium
used only when p is a WebGL plot or the output format is 'webp' or 'svg'. Should be an object of class "rsClientServer" returned by RSelenium::rsDriver.

...
if p is non-WebGL and the output file format is jpeg/png/pdf arguments are passed along to webshot::webshot(). Otherwise, they are ignored.

Details

For SVG plots, a screenshot is taken via webshot::webshot(). Since phantomjs (and hence webshot) does not support WebGL, the RSelenium package is used for exporting WebGL plots.

Author(s)

Carson Sievert

[Package plotly version 4.10.0 Index]