layout {plotly}	R Documentation

Modify the layout of a plotly visualization

Description

Modify the layout of a plotly visualization

Usage

layout(p, ..., data = NULL)

Arguments

p
A plotly object.

...
Arguments to the layout object. For documentation, see https://plotly.com/r/reference/#Layout_and_layout_style_objects

data
A data frame to associate with this layout (optional). If not provided, arguments are evaluated using the data frame in plot_ly().

Author(s)

Carson Sievert

[Package plotly version 4.10.0 Index]