offline {plotly}	R Documentation

Plotly Offline

Description

Deprecated in version 2.0 (offline plots are now the default)

Usage

offline(p, height, width, out_dir, open_browser)

Arguments

p
a plotly object

height
A valid CSS unit. (like "100\ which will be coerced to a string and have "px" appended.

width
A valid CSS unit. (like "100\ which will be coerced to a string and have "px" appended.

out_dir
a directory to place the visualization. If NULL, a temporary directory is used when the offline object is printed.

open_browser
open the visualization after creating it?

Value

a plotly object of class "offline"

Author(s)

Carson Sievert

[Package plotly version 4.10.0 Index]