orca {plotly}	R Documentation

Static image export via orca

Description

The function makes a system call to the orca command-line utility, see the installation instructions here

Usage

orca(p, file = "plot.png", format = tools::file_ext(file), scale = NULL,
  width = NULL, height = NULL, mathjax = FALSE, parallel_limit = NULL,
  verbose = FALSE, debug = FALSE, safe = FALSE)

Arguments

p
a plotly object.

file
output filename.

format
the output format (png, jpeg, webp, svg, pdf, eps).

scale
Sets the image scale. Applies to all output images.

width
Sets the image width. If not set, defaults to layout.width value. Applies to all output images.

height
Sets the image height. If not set, defaults to layout.height value. Applies to all output images.

mathjax
whether or not to include MathJax (required to render TeX). If TRUE, the PLOTLY_MATHJAX_PATH environment variable must be set and point to the location of MathJax (this variable is also used to render TeX in interactive graphs, see config).

parallel_limit
Sets the limit of parallel tasks run.

verbose
Turn on verbose logging on stdout.

debug
Starts app in debug mode and turn on verbose logs on stdout.

safe
Turns on safe mode: where figures likely to make browser window hang during image generating are skipped.

Author(s)

Carson Sievert

Examples

## Not run: 
p <- plot_ly(z = ~volcano) %>% add_surface()
orca(p, "surface-plot.svg")

## End(Not run)

[Package plotly version 4.8.0 Index]