plotly {plotly}	R Documentation

Main interface to plotly

Description

This function is now deprecated. It used to provide a way to store plotly account credentials, but that can now be done with environment variables. For more details and examples, see https://plotly.com/r/getting-started/.

If you're here looking for an intro/overview of the package, see the readme

Usage

plotly(username, key)

Arguments

username
plotly username

key
plotly API key

See Also

ggplotly(), plot_ly(), signup()

[Package plotly version 4.10.0 Index]