plotlyProxy {plotly}	R Documentation

Modify a plotly object inside a shiny app

Description

Modify a plotly object inside a shiny app

Usage

plotlyProxy(
  outputId,
  session = shiny::getDefaultReactiveDomain(),
  deferUntilFlush = TRUE
)

plotlyProxyInvoke(p, method, ...)

Arguments

outputId
single-element character vector indicating the output ID map to modify (if invoked from a Shiny module, the namespace will be added automatically)

session
the Shiny session object to which the map belongs; usually the default value will suffice.

deferUntilFlush
indicates whether actions performed against this instance should be carried out right away, or whether they should be held until after the next time all of the outputs are updated.

p
a plotly proxy object (created with plotlyProxy)

method
a plotlyjs method to invoke. For a list of options, visit https://plotly.com/javascript/plotlyjs-function-reference/

...
unnamed arguments passed onto the plotly.js method

Examples

if (require("shiny") && interactive()) {
  plotly_example("shiny", "proxy_relayout")
  plotly_example("shiny", "proxy_mapbox")
}

[Package plotly version 4.10.0 Index]