plotly_build {plotly}	R Documentation

'Build' (i.e., evaluate) a plotly object

Description

This generic function creates the list object sent to plotly.js for rendering. Using this function can be useful for overriding defaults provided by ggplotly/plot_ly or for debugging rendering errors.

Usage

plotly_build(p, registerFrames = TRUE)

Arguments

p
a ggplot object, or a plotly object, or a list.

registerFrames
should a frame trace attribute be interpreted as frames in an animation?

Examples

p <- plot_ly(economics, x = ~date, y = ~pce)
# the unevaluated plotly object
str(p)
# the evaluated data
str(plotly_build(p)$x$data)

[Package plotly version 4.10.0 Index]